requirements.txt
# Retro Weather Channel Station — Python dependencies
# Install with: pip install -r requirements.txt
#
# Note: tkinter is NOT listed here — it is a system package.
# Install it separately with: sudo apt install python3-tk -y
# Then create your venv with: python3 -m venv venv --system-site-packages
 
# HTTP requests — station data, forecast, alerts, radar, RSS feeds
requests>=2.28.0
 
# Web dashboard
flask>=2.2.5
 
# Weather icon loading and radar image display (touchscreen)
Pillow>=9.0.0
 